Refractored setup code to a container view
