+++ SalesAccountBook/UI/Merch Screen/MerchDetailViewController.swift
@@ -7,29 +7,18 @@ import SnapKit
 class MerchDetailViewController: UIViewController {
     
     // MARK: - Variables
-    let containerView: UIView
-    let merchPic: MerchIconView
-    let merchName: TitleWithTextField
-    let merchPrice: TitleWithTextField
-    let merchQty: TitleWithTextField
-    let merchRemark: TitleWithTextField
-    let actionType: DetailsViewActionType
+    let containerView: MerchFieldsContainer
     
+    let actionType: DetailsViewActionType
     var currentMerchName: String?
     weak var inventory: Inventory?
     var onSelectRowDelegate: ((String) -> Void)?
     
     // MARK: - Initializion
     init(config: MerchDetailsConfigurator) {
-        self.containerView = UIView()
-        self.merchPic = MerchIconView()
-        self.merchName = TitleWithTextField(title: NSLocalizedString("MerchName", comment: "Name of product."))
-        self.merchPrice = TitleWithTextField(title: NSLocalizedString("MerchPrice", comment: "Price of product."))
-        self.merchQty = TitleWithTextField(title: NSLocalizedString("MerchQty", comment: "Quantity of product."))
-        self.merchRemark = TitleWithTextField(title: NSLocalizedString("MerchRemark", comment: "Remark of product."))
+        self.containerView = MerchFieldsContainer()
         
         self.actionType = config.action
-        
         self.currentMerchName = config.merchName
         self.inventory = config.inventory
         self.onSelectRowDelegate = config.onSelectRow
@@ -59,34 +48,22 @@ class MerchDetailViewController: UIViewController {
         self.setup()
     }
     
-    private func setupPlaceholders() {
-        self.merchName.textField.placeholder = NSLocalizedString("Required(Input)", comment: "Must input.")
-        self.merchPrice.textField.placeholder = NSLocalizedString("Optional(Input)", comment: "Can leave blank.")
-        self.merchQty.textField.placeholder = NSLocalizedString("Optional(Input)", comment: "Can leave blank.")
-        self.merchRemark.textField.placeholder = NSLocalizedString("Optional(Input)", comment: "Can leave blank.")
-    }
-    
     private func prefillFieldsForEdit() {
         guard let merchDetails = self.inventory?.getMerch(name: self.currentMerchName ?? "") else {
             fatalError()
         }
         
         if let img = merchDetails.image {
-            self.merchPic.iconImage.image = img
+            self.containerView.merchPic.iconImage.image = img
         }
-        self.merchName.textField.text = merchDetails.name
-        self.merchPrice.textField.text = String(merchDetails.price)
-        self.merchQty.textField.text = String(merchDetails.qty)
-        self.merchRemark.textField.text = merchDetails.remark
+        self.containerView.merchName.textField.text = merchDetails.name
+        self.containerView.merchPrice.textField.text = String(merchDetails.price)
+        self.containerView.merchQty.textField.text = String(merchDetails.qty)
+        self.containerView.merchRemark.textField.text = merchDetails.remark
     }
     
     private func setup() {
         self.view.backgroundColor = .groupTableViewBackground
-        self.setupPlaceholders()
-        
-        if self.actionType == .edit {
-            self.prefillFieldsForEdit()
-        }
         
         self.view.addSubview(self.containerView)
         self.containerView.snp.makeConstraints { make in
@@ -95,70 +72,15 @@ class MerchDetailViewController: UIViewController {
         }
         self.containerView.backgroundColor = .white
         
-        self.containerView.addSubview(self.merchPic)
-        self.merchPic.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.left.equalToSuperview()
-            make.right.equalToSuperview()
-            make.height.equalTo(self.view).dividedBy(5)
-        }
-        self.merchPic.addLine(position: .LINE_POSITION_BOTTOM, color: .groupTableViewBackground, width: 1)
-        self.merchPic.backgroundColor = .white
-        DispatchQueue.main.async {
-            self.merchPic.iconImage.clipsToBounds = true
-            self.merchPic.iconImage.layer.cornerRadius = self.merchPic.iconImage.frame.width / 2
+        self.containerView.setup()
+        
+        if self.actionType == .edit {
+            self.prefillFieldsForEdit()
         }
         
         let tapGest = UITapGestureRecognizer(target: self, action: #selector(self.showImageUploadOption))
-        self.merchPic.addGestureRecognizer(tapGest)
-        self.merchPic.isUserInteractionEnabled = true
-        
-        self.containerView.addSubview(self.merchName)
-        self.merchName.snp.makeConstraints { make in
-            make.top.equalTo(self.merchPic.snp.bottom)
-            make.left.equalToSuperview().offset(Constants.UI.Spacing.Small)
-            make.right.equalToSuperview().offset(-Constants.UI.Spacing.Small)
-            make.height.equalTo(44)
-        }
-        self.merchName.textField.clearButtonMode = .whileEditing
-        self.merchName.backgroundColor = .white
-        self.merchName.addLine(position: .LINE_POSITION_BOTTOM, color: .groupTableViewBackground, width: 1)
-        
-        self.containerView.addSubview(self.merchPrice)
-        self.merchPrice.snp.makeConstraints { make in
-            make.top.equalTo(self.merchName.snp.bottom)
-            make.left.equalToSuperview().offset(Constants.UI.Spacing.Small)
-            make.right.equalToSuperview().offset(-Constants.UI.Spacing.Small)
-            make.height.equalTo(44)
-        }
-        self.merchPrice.textField.clearButtonMode = .whileEditing
-        self.merchPrice.textField.keyboardType = .numberPad
-        self.merchPrice.backgroundColor = .white
-        self.merchPrice.addLine(position: .LINE_POSITION_BOTTOM, color: .groupTableViewBackground, width: 1)
-        
-        self.containerView.addSubview(self.merchQty)
-        self.merchQty.snp.makeConstraints { make in
-            make.top.equalTo(self.merchPrice.snp.bottom)
-            make.left.equalToSuperview().offset(Constants.UI.Spacing.Small)
-            make.right.equalToSuperview().offset(-Constants.UI.Spacing.Small)
-            make.height.equalTo(44)
-        }
-        self.merchQty.textField.clearButtonMode = .whileEditing
-        self.merchQty.textField.keyboardType = .numberPad
-        self.merchQty.backgroundColor = .white
-        self.merchQty.addLine(position: .LINE_POSITION_BOTTOM, color: .groupTableViewBackground, width: 1)
-        
-        self.containerView.addSubview(self.merchRemark)
-        self.merchRemark.snp.makeConstraints { make in
-            make.top.equalTo(self.merchQty.snp.bottom)
-            make.left.equalToSuperview().offset(Constants.UI.Spacing.Small)
-            make.right.equalToSuperview().offset(-Constants.UI.Spacing.Small)
-            make.height.equalTo(44)
-            make.bottom.equalToSuperview()
-        }
-        self.merchRemark.textField.clearButtonMode = .whileEditing
-        self.merchRemark.backgroundColor = .white
-        self.merchRemark.addLine(position: .LINE_POSITION_BOTTOM, color: .groupTableViewBackground, width: 1)
+        self.containerView.merchPic.addGestureRecognizer(tapGest)
+        self.containerView.merchPic.isUserInteractionEnabled = true
     }
     
     
@@ -168,7 +90,7 @@ class MerchDetailViewController: UIViewController {
         let handler: (UIAlertAction) -> Void = { [weak self] alert in
             guard let `self` = self else { return }
             
-            guard let merchNameText = self.merchName.textField.text, self.merchName.textField.text != "" else {
+            guard let merchNameText = self.containerView.merchName.textField.text, self.containerView.merchName.textField.text != "" else {
                 self.promptEmptyMerchNameError()
                 return
             }
@@ -210,10 +132,10 @@ class MerchDetailViewController: UIViewController {
     }
     
     private func makeMerchDetails(name: String) -> MerchDetails {
-        let parsedPrice = Double(self.merchPrice.textField.text ?? "") ?? 0.0
-        let parsedQty = Int(self.merchQty.textField.text ?? "") ?? 0
+        let parsedPrice = Double(self.containerView.merchPrice.textField.text ?? "") ?? 0.0
+        let parsedQty = Int(self.containerView.merchQty.textField.text ?? "") ?? 0
         
-        return (name: name, price: parsedPrice, qty: parsedQty, remark: (self.merchRemark.textField.text) ?? "", image: self.merchPic.iconImage.image)
+        return (name: name, price: parsedPrice, qty: parsedQty, remark: (self.containerView.merchRemark.textField.text) ?? "", image: self.containerView.merchPic.iconImage.image)
     }
     
     private func addMerch(merchDetails: MerchDetails) {
@@ -250,7 +172,7 @@ class MerchDetailViewController: UIViewController {
         let errorMessage = NSLocalizedString("ErrorMerchInputEmpty", comment: "Error Message - Merch name text field .")
         self.present(UIAlertController.makeError(message: errorMessage), animated: true, completion: nil)
         
-        self.merchName.textField.attributedPlaceholder = NSAttributedString(
+        self.containerView.merchName.textField.attributedPlaceholder = NSAttributedString(
             string: NSLocalizedString("Required(Input)", comment: "Must input."),
             attributes: [NSAttributedString.Key.foregroundColor: UIColor.red.withAlphaComponent(0.5)])
     }
@@ -278,7 +200,7 @@ extension MerchDetailViewController: UIImagePickerControllerDelegate, UINavigati
         
         actionSheet.addAction(UIAlertAction(title: NSLocalizedString("Cancel", comment: "Decide an event will not take place."), style: .cancel, handler: nil))
         
-        resetBtn.isEnabled = !(self.merchPic.iconImage.image == #imageLiteral(resourceName: "MerchDefault"))
+        resetBtn.isEnabled = !(self.containerView.merchPic.iconImage.image == #imageLiteral(resourceName: "MerchDefault"))
         
         self.present(actionSheet, animated: true, completion: nil)
     }
@@ -304,12 +226,12 @@ extension MerchDetailViewController: UIImagePickerControllerDelegate, UINavigati
     }
     
     private func resetIconDefault() {
-        self.merchPic.iconImage.image = #imageLiteral(resourceName: "MerchDefault")
+        self.containerView.merchPic.iconImage.image = #imageLiteral(resourceName: "MerchDefault")
     }
     
     func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
         if let img = info[UIImagePickerController.InfoKey.editedImage] as? UIImage {
-            self.merchPic.iconImage.image = img
+            self.containerView.merchPic.iconImage.image = img
         }
         self.dismiss(animated: true, completion: nil)
     }
